Separate tensorflow model runner code from serving framework as a sub-library.
This allows extraneous dependencies to be dropped from cases that don't require them.

PiperOrigin-RevId: 759755574
