+++ python/serving/serving_framework/tensorflow/requirements.in
@@ -1,0 +1,3 @@
+tensorflow~=2.18.0
+tensorflow-serving-api~=2.18.0
+tensorflow-io-gcs-filesystem>=0.23.1
